fix: move components to components-v2
