+++ apps/gitness/src/components-v2/standalone/side-bar.tsx
@@ -5,13 +5,13 @@ import { ManageNavigation, MenuGroupType, MenuGroupTypes, NavbarItemType, Sideba
 import { useTranslation } from '@harnessio/ui/context'
 import { SidebarView } from '@harnessio/ui/views'
 
-import { useNav } from '../../components/stores/recent-pinned-nav-links.store'
 import { getNavbarMenuData } from '../../data/navbar-menu-data'
 import { useAppContext } from '../../framework/context/AppContext'
 import { useRoutes } from '../../framework/context/NavigationContext'
 import { useSelectedSpaceId } from '../../framework/hooks/useSelectedSpaceId'
 import { useTranslationStore } from '../../i18n/stores/i18n-store'
 import { PathParams } from '../../RouteDefinitions'
+import { useNav } from '../stores/recent-pinned-nav-links.store'
 
 const AppSideBar: FC<{ children: React.ReactNode }> = ({ children }) => {
   const { t } = useTranslation()
