+++ apps/gitness/src/pages-v2/repo/repo-execution-list.tsx
@@ -7,7 +7,7 @@ import { ListExecutionsOkResponse, TypesExecution, useListExecutionsQuery } from
 import { IconV2 } from '@harnessio/ui/components'
 import { ExecutionListPage, IExecution } from '@harnessio/ui/views'
 
-import { LinkComponent } from '../../components/LinkComponent'
+import { LinkComponent } from '../../components-v2/LinkComponent'
 import { useGetRepoRef } from '../../framework/hooks/useGetRepoPath'
 import { parseAsInteger, useQueryState } from '../../framework/hooks/useQueryState'
 import { PathParams } from '../../RouteDefinitions'
